fix non-breaking urls in sidenote
I noticed on my own site that sidenotes and marginnotes with links were a mess. the links didn't break into the container they were supposed to stay in unless there was a hyphen present. This cleans that up. 

I thought that you might be able to use it as well. I put it where I thought it made the most sense. 

Reference:
https://css-tricks.com/snippets/css/prevent-long-urls-from-breaking-out-of-container/